refac: admin - 쿠폰탭 분리
